Renamed getCoords module file
