+++ src/shared/lib/dom/get-coords.ts
