+++ src/shared/lib/dom/index.ts
@@ -1,1 +1,1 @@
-export { default as getCoords } from './getCoords';
+export { default as getCoords } from './get-coords';
